refactor todo update logic
